Add keyvalut-appinsights-docker etc
