--- Altinn.Platform.DialogportenAdapter.sln
@@ -1,41 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "DialogportenAdapter", "src\DialogportenAdapter\DialogportenAdapter.csproj", "{B1D4ADAD-3176-4C3E-A0D2-E313792587CA}"
-EndProject
-Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution Items", "{530DD6C7-2F29-4B27-9FEE-0F2612F5391B}"
-	ProjectSection(SolutionItems) = preProject
-		docker-compose.yml = docker-compose.yml
-	EndProjectSection
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "DialogportenAdapter.Interface", "src\DialogportenAdapter.Interface\DialogportenAdapter.Interface.csproj", "{62EFCB3C-3C4D-418B-A4DD-DC1011AF8469}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Altinn.Platform.DialogportenAdapter.UnitTest", "test\UnitTest\Altinn.Platform.DialogportenAdapter.UnitTest.csproj", "{097FDDA8-CBDE-4C68-92C1-DBD9A36FD6D8}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Altinn.Platform.DialogportenAdapter.WebApi", "src\Altinn.Platform.DialogportenAdapter.WebApi\Altinn.Platform.DialogportenAdapter.WebApi.csproj", "{3FB848BE-8FD5-4DA2-BC83-EE7CBA8359E3}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{B1D4ADAD-3176-4C3E-A0D2-E313792587CA}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{B1D4ADAD-3176-4C3E-A0D2-E313792587CA}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{B1D4ADAD-3176-4C3E-A0D2-E313792587CA}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{B1D4ADAD-3176-4C3E-A0D2-E313792587CA}.Release|Any CPU.Build.0 = Release|Any CPU
-		{62EFCB3C-3C4D-418B-A4DD-DC1011AF8469}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{62EFCB3C-3C4D-418B-A4DD-DC1011AF8469}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{62EFCB3C-3C4D-418B-A4DD-DC1011AF8469}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{62EFCB3C-3C4D-418B-A4DD-DC1011AF8469}.Release|Any CPU.Build.0 = Release|Any CPU
-		{097FDDA8-CBDE-4C68-92C1-DBD9A36FD6D8}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{097FDDA8-CBDE-4C68-92C1-DBD9A36FD6D8}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{097FDDA8-CBDE-4C68-92C1-DBD9A36FD6D8}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{097FDDA8-CBDE-4C68-92C1-DBD9A36FD6D8}.Release|Any CPU.Build.0 = Release|Any CPU
-		{3FB848BE-8FD5-4DA2-BC83-EE7CBA8359E3}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{3FB848BE-8FD5-4DA2-BC83-EE7CBA8359E3}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{3FB848BE-8FD5-4DA2-BC83-EE7CBA8359E3}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{3FB848BE-8FD5-4DA2-BC83-EE7CBA8359E3}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(NestedProjects) = preSolution
-	EndGlobalSection
-EndGlobal
