+++ docker-compose.yml
@@ -1,6 +1,8 @@
 ﻿services:
   dialogportenadapter:
     image: dialogportenadapter
+    ports:
+      - 8080:5011
     build:
       context: .
-      dockerfile: src/DialogportenAdapter/Dockerfile
+      dockerfile: Dockerfile
