+++ src/Altinn.DialogportenAdapter.WebApi/appsettings.Development.json
@@ -1,0 +1,28 @@
+{
+  "Logging": {
+    "LogLevel": {
+      "Default": "Information",
+      "Microsoft.AspNetCore": "Warning"
+    }
+  },
+  "Infrastructure": {
+    "Maskinporten": {
+      "Environment": "test",
+      "TokenExchangeEnvironment": "tt02",
+      "Scope": "digdir:dialogporten.serviceprovider.admin altinn:serviceowner/instances.write digdir:dialogporten.serviceprovider altinn:serviceowner/instances.read",
+      "ClientId": "TODO: Add to local secrets",
+      "EncodedJwk": "TODO: Add to local secrets",
+      "ExhangeToAltinnToken": true,
+      "EnableDebugLogging": false
+    },
+    "Altinn": {
+      "BaseUri": "https://tt02.altinn.no"
+    },
+    "Dialogporten": {
+      "BaseUri": "https://altinn-dev-api.azure-api.net/dialogporten"
+    },
+    "Adapter": {
+      "BaseUri": "https://localhost:5011"
+    }
+  }
+}
