+++ src/Altinn.DialogportenAdapter.WebApi/appsettings.json
@@ -1,0 +1,29 @@
+{
+  "Logging": {
+    "LogLevel": {
+      "Default": "Information",
+      "Microsoft.AspNetCore": "Warning"
+    }
+  },
+  "DialogportenAdapter": {
+    "Maskinporten": {
+      "Environment": "test",
+      "TokenExchangeEnvironment": "tt02",
+      "Scope": "digdir:dialogporten.serviceprovider.admin altinn:serviceowner/instances.write digdir:dialogporten.serviceprovider altinn:serviceowner/instances.read",
+      "ClientId": "TODO: Add to local secrets",
+      "EncodedJwk": "TODO: Add to local secrets",
+      "ExhangeToAltinnToken": true,
+      "EnableDebugLogging": false
+    },
+    "Altinn": {
+      "BaseUri": "https://tt02.altinn.no",
+      "SubscriptionKey": "SubscriptionKey"
+    },
+    "Dialogporten": {
+      "BaseUri": "https://altinn-dev-api.azure-api.net/dialogporten"
+    },
+    "Adapter": {
+      "BaseUri": "https://localhost:5011"
+    }
+  }
+}
