+++ src/Altinn.Platform.DialogportenAdapter.WebApi/Altinn.DialogportenAdapter.WebApi.csproj
@@ -1,0 +1,1 @@
+﻿
