--- src/Altinn.Platform.DialogportenAdapter.WebApi/Altinn.Platform.DialogportenAdapter.WebApi.csproj
@@ -1,17 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk.Web">
-
-    <PropertyGroup>
-        <TargetFramework>net9.0</TargetFramework>
-        <Nullable>enable</Nullable>
-        <ImplicitUsings>enable</ImplicitUsings>
-        <UserSecretsId>f15beaf9-69f4-454a-8605-6586ecad451d</UserSecretsId>
-    </PropertyGroup>
-
-    <ItemGroup>
-        <PackageReference Include="Altinn.ApiClients.Maskinporten" Version="9.2.1" />
-        <PackageReference Include="Altinn.Platform.Storage.Interface" Version="4.0.4" />
-        <PackageReference Include="Microsoft.AspNetCore.OpenApi" Version="9.0.0"/>
-        <PackageReference Include="Refit.HttpClientFactory" Version="8.0.0" />
-    </ItemGroup>
-
-</Project>
